Housekeeping - modularization of test cases
